ir: send sec-dec-cent instead of dec-cent-mil as value
Because otherwise numbers can be similar when harvester is sending at 100% perfect rate signal. By sending seconds first the variability is ensured. And by sending dec and cent, the randomization is enough to display a real change. The fourth number (miiliseconds) is removed because for human eyes, spot and process 3 numbers is much better than 4.