serial: add stream/emmitters uncaught errors handler
First of all, this commit is to HOT DEBUG a problem spotted sometimes. Even if not solved, the handler will call SIGINT so the daemon will restart the app

The important thing here is to LOG that a problem was raised because currently the app stops sending harvesters id after a while running. We need to check if the stream error is not killing properly the app (no SIGINT) of if there is any kind of BUG in code (either mine or in the serial-port or in the node version)