mqtt: MAJOR, process all events to detect not connection.
IMPORTANT, if not set, a memory leak appears in the app because the client is always in the state connected but not connected, and no desconection occurs for a whiel, so if multimple mqtt calls, multiple "client" objects alive.
Anoted IMPORTANT thing is that the mqtt librery DOES NOT reject when the server address is invalid or unreachable, ant the only event to detect that case is the "disconnect" event.

Since this app is intended to connect, send and forget, the client.end() is called after some events.

Also, the disconnect and other events hasve been implemented to log in production for problems because one of the issues was that we weren't aware on the problems because nothing appeared in the journalctl, and the probm was that System Managers at Tecnalia changed the output ip, and that new IP was not in the INDRA's whitelist.

So a ton of time to detect that and later another ton of time to realize that the eventemiiter for mqtt was not as expected (if problem, rise error or execption at least!)