mqtt: test, check INDRA messages and also check our autogenerated mqtt messages
Comprehensive tests because of the object, map reordering of keys problem.

In the message.js, solved because the 5750 key is now at the bottom so input string and deserialization into object have the same key ordering

For INDRA messages, configure the variables in the test and run in. That test applies the patch for reordering keys.

For Tecnalia messages, just check that the message is being generated ok