clean: mqtt subproject only with min documentation
The other projects do require only a SMOOL SIB (broker) to run, and some 433Mhz devices emulating the harvesters devices. In the mqtt project, the requirements are external CMW and SFTM not provided here, so better to state in the inner documentation than provide an non-updated implementation. The CMW is still in constant developments so old MQTT messages provided by this repository may not work.